messaging_service/requirements.txt
aiogram>=3.6,<4
aiohttp>=3.9
pyyaml==6.0.1
loguru==0.7.2
python-dotenv>=1.0.0
injector==0.21.0
pydantic>=2.0.0
qrcode==7.4.2
Pillow>=10.0.0
paho-mqtt>=2.1.0
cairosvg>=2.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Analytics storage
aiosqlite>=0.19

# CLI
typer[all]>=0.12.0
rich>=13.0.0
